Move after_request to rate_limiter
