+++ webserver/rate_limiter.py
@@ -28,6 +28,16 @@ RATELIMIT_WINDOW_DEFAULT = 10
 RATELIMIT_REFRESH = 60 # in seconds
 RATELIMIT_TIMEOUT = "rate_limits_timeout"
 
+# Add rate limit headers to responses
+def inject_x_rate_headers(response):
+    limit = get_view_rate_limit()
+    if limit and limit.send_x_headers:
+        h = response.headers
+        h.add('X-RateLimit-Remaining', str(limit.remaining))
+        h.add('X-RateLimit-Limit', str(limit.limit))
+        h.add('X-RateLimit-Reset', str(limit.reset))
+    return response
+
 class RateLimit(object):
     
     # From the docs:
